Consolidate B-tree code into single file
